fix(dependency installer): fix raising PermissionError when installing dependencies on Windows
The error is raised when the program is trying to unzip the downloaded tarballs into the %ProgramFiles% directory, which requires administrative privileges. Obtaining administrative privileges is difficult due to the code organization of this program. This commit changes the target directory of storing dependencies from %ProgramFiles% to the accessible %AppData% directory of the logged-in user.

A drawback of this change is that dependencies are installed only for
that single user and thus could actually be repeatedly installed by
other users on the same computer (https://learn.microsoft.com/en-us/answers/questions/584279/whats-the-idea-behind-the-name-of-the-appdata-fold).